+++ typescript/agentkit/src/wallet-providers/index.ts
@@ -2,3 +2,5 @@ export * from "./walletProvider";
 export * from "./evmWalletProvider";
 export * from "./viemWalletProvider";
 export * from "./cdpWalletProvider";
+export * from "./svmWalletProvider";
+export * from "./solanaKeypairWalletProvider";